feat: Admins can now edit Category names (#12)
Admin users now have the ability to modify existing Category names. This feature allows for improved content organization and the flexibility to rephrase Category names as needed. The user is presented with an edit form upon selecting the option to edit a Category. Upon saving the changes, the user is redirected back to the updated Category list. In case of deciding not to proceed with the edit, a 'Cancel' button is provided to return the user to the Category list.

Note: As of now, all users are treated as admin users. User permissions will be strictly enforced in upcoming updates.

f